+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2.0.0
+
+- Add a `beforeEach` function that can be used on a per-spec basis. Thanks to [PatrickBRT](https://github.com/PatrickBRT) whose work inspired our approach!
+- **BREAKING** Clear AsyncStorage and re-render the app before each test runs.
+- Cavy no longer resets your app at the end of the test suite.
+
 # 1.1.0
 
 - Un-deprecate `wrap` (was deprecated in 0.6.0) and rewrite it using React